Flip out and rewrite everything
- Place all code in internal_ions/ so that imports work properly
- Create SpectrumFile and PSMList and cache them, use session_state less
- Fix use of constants
