+++ internal_ions/fragannot/fragannot_call.py
@@ -2,11 +2,11 @@ from .fragannot_numba import FragannotNumba
 
 from typing import Dict
 from typing import List
-from typing import BinaryIO
+from ..util.spectrumio import SpectrumFile
 from psm_utils.psm_list import PSMList
 
 
-def fragannot_call(spectrum_file: BinaryIO,
+def fragannot_call(spectrum_file: SpectrumFile,
                    psms: PSMList,
                    tolerance: float,
                    fragment_types: List[str],
