--- util/constants.py
@@ -1,6 +1,0 @@
-from psm_utils.io import FILETYPES as _ftypes
-SUPPORTED_FILETYPES = list(_ftypes.keys())
-
-REPO_NAME = "internal_ions"
-DIV_COLOR = "rainbow"
-FRAGANNOT_ION_NAMES = ["a", "b", "c", "cdot", "c-1", "c+1", "x", "y", "zdot", "z+1", "z+2", "z+3"]
